backend/services/ride-service/requirements.txt
# FastAPI Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# HTTP Client (for calling user-service)
httpx>=0.27.0

# Security (for verifying tokens if needed, though usually gateway handles this)
python-jose[cryptography]==3.3.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
